refactor: 💡 update Entity files
Entity파일 및 기타 DB 테이블에 관련된 코드 수정 및 테스트

+++ yogidice/src/main/java/com/specialization/yogidice/dto/request/category/MechanismGroupUpdateRequest.java
@@ -11,5 +11,5 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 public class MechanismGroupUpdateRequest {
     @NotNull
-    private Long mechanismId;
+    private Byte mechanismId;
 }
